refactor(dashboard): update SelectedAssetLite imports to reference types/disposals
