+++ src/components/features/dashboard/global-admin/global-admin-dashboard-view.tsx
@@ -5,7 +5,8 @@ import { Tabs, TabsList, TabsTrigger, TabsContent } from "@/components/ui/tabs"
 import { DataTable } from "@/components/shared/data-table"
 import { cn } from "@/lib/utils"
 import { TYPOGRAPHY_CLASSNAMES } from "@/components/shared/typography"
-import { DisposeAssetsRequestDialog, type SelectedAssetLite } from "@/components/features/disposals/dispose-assets-request-dialog"
+import { DisposeAssetsRequestDialog } from "@/components/features/disposals/dispose-assets-request-dialog"
+import type { SelectedAssetLite } from "@/types/disposals"
 import { tiqriToast } from "@/components/shared/sonner"
 import { sendAssignmentReminderAction } from "@/actions/assignments"
 import { KpiMetricsRow } from "../shared/kpi-metrics-row"
